+++ library/src/types/index.ts
@@ -1,4 +1,3 @@
-export * from './hash.ts';
 export * from './issues.ts';
 export * from './other.ts';
 export * from './pipe.ts';