+++ library/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the library will be documented in this file.
 
 ## vX.X.X (Month DD, YYYY)
 
-- Add `creditCard`, `hex` and `hexColor` validation function (pull request #292, #304, #308)
+- Add `creditCard`, `hash`, `hex` and `hexColor` validation function (pull request #292, #304, #307, #308)
 
 ## v0.24.1 (December 11, 2023)